Add endpoints registers

+++ src/main/java/com/ajsw/javausersservice/repositories/interfaces/IInstructorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IInstructorRepository extends JpaRepository<Instructor, Long> {
+public interface IInstructorRepository extends JpaRepository<Instructor, Integer> {
 }
